# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ChannelType:
    enum:
    - email
    - slack
    - teams
    - webhook
    type: string
    x-enum-varnames:
    - ChannelTypeEmail
    - ChannelTypeSlack
    - ChannelTypeTeams
    - ChannelTypeWebhook
...
  domain.CreateNotificationChannelRequest:
    properties:
      channel_type:
        $ref: '#/definitions/domain.ChannelType'
      config:
        type: object
      is_enabled:
        type: boolean
      name:
        type: string
    required:
    - channel_type
    - config
    - name
    type: object
  domain.CreateUserNotificationPreferenceRequest:
    properties:
      channel_id:
        type: string
      dnd_enabled:
        type: boolean
      dnd_end_time:
        type: string
      dnd_start_time:
        type: string
      is_enabled:
        type: boolean
      min_priority:
        type: string
    required:
    - channel_id
    type: object
...
  domain.NotificationChannel:
    properties:
      channel_type:
        $ref: '#/definitions/domain.ChannelType'
      config:
        type: object
      created_at:
        type: string
      id:
        type: string
      is_enabled:
        type: boolean
      name:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
    type: object
  domain.NotificationLog:
    properties:
      alert_id:
        type: string
      channel_id:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      message:
        type: string
      organization_id:
        type: string
      recipient:
        type: string
      sent_at:
        type: string
      status:
        $ref: '#/definitions/domain.NotificationStatus'
      subject:
        type: string
      user_id:
        type: string
    type: object
  domain.NotificationStatus:
    enum:
    - pending
    - sent
    - failed
    type: string
    x-enum-varnames:
    - NotificationStatusPending
    - NotificationStatusSent
    - NotificationStatusFailed
...
  domain.SendNotificationRequest:
    properties:
      alert_id:
        type: string
      channel_id:
        type: string
      message:
        type: string
      recipient:
        type: string
      subject:
        type: string
      user_id:
        type: string
    required:
    - channel_id
    - message
    - recipient
    type: object
...
  domain.UpdateNotificationChannelRequest:
    properties:
      channel_type:
        $ref: '#/definitions/domain.ChannelType'
      config:
        type: object
      is_enabled:
        type: boolean
      name:
        type: string
    type: object
  domain.UpdateUserNotificationPreferenceRequest:
    properties:
      dnd_enabled:
        type: boolean
      dnd_end_time:
        type: string
      dnd_start_time:
        type: string
      is_enabled:
        type: boolean
      min_priority:
        type: string
    type: object
...
  domain.UserNotificationPreference:
    properties:
      channel_id:
        type: string
      created_at:
        type: string
      dnd_enabled:
        type: boolean
      dnd_end_time:
        type: string
      dnd_start_time:
        type: string
      id:
        type: string
      is_enabled:
        type: boolean
      min_priority:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pulsar.io
    name: API Support
    url: https://github.com/nmn3m/pulsar
  description: Pulsar is an incident management and alerting platform. This API provides
    endpoints for managing alerts, incidents, teams, schedules, escalation policies,
    notifications, and webhooks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pulsar API
  version: "1.0"
paths:
...
  /notifications/channels:
    get:
      consumes:
      - application/json
      description: Lists all notification channels for the organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notification channels
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Creates a new notification channel for the organization
      parameters:
      - description: Channel creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.NotificationChannel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a notification channel
      tags:
      - Notifications
  /notifications/channels/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a notification channel by its ID
      parameters:
      - description: Channel ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a notification channel
      tags:
      - Notifications
    get:
      consumes:
      - application/json
      description: Retrieves a notification channel by its ID
      parameters:
      - description: Channel ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.NotificationChannel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a notification channel
      tags:
      - Notifications
    patch:
      consumes:
      - application/json
      description: Updates an existing notification channel by its ID
      parameters:
      - description: Channel ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Channel update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.NotificationChannel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a notification channel
      tags:
      - Notifications
  /notifications/logs:
    get:
      consumes:
      - application/json
      description: Lists all notification logs for the organization with pagination
      parameters:
      - default: 50
        description: Number of logs to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of logs to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notification logs
      tags:
      - Notifications
  /notifications/logs/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a notification log entry by its ID
      parameters:
      - description: Log ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.NotificationLog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a notification log
      tags:
      - Notifications
  /notifications/logs/alert/{alertId}:
    get:
      consumes:
      - application/json
      description: Lists all notification logs associated with a specific alert
      parameters:
      - description: Alert ID
        format: uuid
        in: path
        name: alertId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notification logs by alert
      tags:
      - Notifications
  /notifications/logs/user/me:
    get:
      consumes:
      - application/json
      description: Lists all notification logs for the authenticated user with pagination
      parameters:
      - default: 50
        description: Number of logs to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of logs to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notification logs for the current user
      tags:
      - Notifications
  /notifications/preferences:
    get:
      consumes:
      - application/json
      description: Lists all notification preferences for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List user notification preferences
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Creates a new notification preference for the authenticated user
      parameters:
      - description: Preference creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUserNotificationPreferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.UserNotificationPreference'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a user notification preference
      tags:
      - Notifications
  /notifications/preferences/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a notification preference by its ID
      parameters:
      - description: Preference ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user notification preference
      tags:
      - Notifications
    get:
      consumes:
      - application/json
      description: Retrieves a notification preference by its ID
      parameters:
      - description: Preference ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserNotificationPreference'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a user notification preference
      tags:
      - Notifications
    patch:
      consumes:
      - application/json
      description: Updates an existing notification preference by its ID
      parameters:
      - description: Preference ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Preference update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserNotificationPreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserNotificationPreference'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user notification preference
      tags:
      - Notifications
  /notifications/send:
    post:
      consumes:
      - application/json
      description: Sends a notification through the specified channel
      parameters:
      - description: Notification send request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.NotificationLog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send a notification
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"